s3: Fix bug 7398 -- rename problems with full_audit

+++ source3/modules/vfs_full_audit.c
@@ -1094,7 +1094,7 @@ static int smb_full_audit_lstat(vfs_handle_struct *handle,
 static uint64_t smb_full_audit_get_alloc_size(vfs_handle_struct *handle,
 		       files_struct *fsp, const SMB_STRUCT_STAT *sbuf)
 {
-	int result;
+	uint64_t result;
 
 	result = SMB_VFS_NEXT_GET_ALLOC_SIZE(handle, fsp, sbuf);
